Langgraph Frontend Logs appear without overloading react state
